+++ Runtimes/Core/Concurrency/CMakeLists.txt
@@ -16,7 +16,6 @@ add_library(swift_Concurrency
   ExecutorImpl.cpp
   ExecutorChecks.cpp
   GlobalExecutor.cpp
-  Setup.cpp
   Task.cpp
   TaskAlloc.cpp
   TaskGroup.cpp
